feat: update styling
